requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
beautifulsoup4==4.11.1
certifi==2022.6.15
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-query-counter==0.2.1
django-redis==5.2.0
django-rest-fuzzysearch==0.5.1
django-rest-multiple-models==2.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
html5lib==1.1
httpie==3.2.1
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
iniconfig==1.1.1
model-bakery==1.7.0
msgpack==1.0.5
packaging==21.3
pluggy==1.0.0
psycopg2-binary
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyngrok==6.0.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
redis==4.5.5
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tabulate==0.9.0
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2022.1
urllib3==1.26.11
webencodings==0.5.1
wrapt==1.14.1
zope.interface==6.0
nepali-datetime==1.0.8.1